fix: UaF in boot-manifest-hash fallback
CFDataGetBytePtr doesn't return a copy. Somehow it happened to work fine
up until iOS 16.3.1, but on iOS 16.4+ the early CFRelease causes the
result to be all zeroes.

This isn't critical as it's just a fallback if it fails to get
root-snapshot-name and that still works fine, but it's worth fixing
anyway.